requirements.txt
django
djangorestframework
django-grappelli
django-filter
django-filebrowser
pillow
mutagen
graphene
graphene-django